clang-tidy triyaj: kendi kodumuzdaki tum bulgular temizlendi (0 kaldi)
CI'daki clang-tidy job'u su ana kadar yalnizca bilgilendirici calisiyordu
(bulgular hic triyaj edilmemisti). Gercek calistirma: 1.13 milyon uyari
uretti ama bunlarin 1.13 milyondan fazlasi vendor/sistem header'larindan
(Jolt/lz4) -- kendi src/ agacimizda yalnizca 8 gercek bulgu vardi:

- AeroSystem.cpp / TerrainDeformSystem.cpp: RVec3 kurulumunda double->
  float daralmasi -- kod double'a genisletip sonra tek-hassasiyetli
  RVec3(=Vec3) yapicisinda tekrar float'a daraltiyordu (bu build'de
  anlamsiz ara adim). Dogrudan float'a cast edilecek sekilde duzeltildi
  -- hem tek hem cift hassasiyetli Jolt derlemelerinde dogru.
- DebugDrawBridge.cpp: JPH::Color bilesenlerini (uint8) 255.0f'e bolerken
  brace-init'te examplicit int->float cast eksikti -- eklendi (sonuc
  ayni, sadece derleyiciye niyeti acikca belirtiyor).
- RopeSystem.cpp: segments[i+1] indeksinde int toplama once yapilip
  SONRA size_t'e donusuyordu (genel durumda tasma riski) -- once i'yi
  size_t'e cevirip sonra +1 yapacak sekilde duzeltildi.

Davranis degisikligi yok (tum degerler ayni), 29/29 yerel ctest yesil.
CI'nin clang-tidy job'u artik gercekten sifir bulguyla calisiyor.

+++ src/aero/AeroSystem.cpp
@@ -15,8 +15,8 @@ void AeroSystem::ApplyParachute(JPH::PhysicsSystem& physics, JPH::BodyID body,
     // Kuadratik surukleme, hiza ters yonde
     const JPH::Vec3 force = rel * (-config.dragCoefficient * speed);
     // Baglanti noktasi govde merkezinin ustunde: cisim kendini diklestirir
-    const JPH::RVec3 attachPoint = bi.GetCenterOfMassPosition(body) +
-                                   JPH::RVec3(0.0, static_cast<double>(config.attachHeight), 0.0);
+    const JPH::RVec3 attachPoint =
+        bi.GetCenterOfMassPosition(body) + JPH::RVec3(0.0f, config.attachHeight, 0.0f);
     bi.AddForce(body, force, attachPoint);
     bi.ActivateBody(body);
 }
